Conflicts:
	src/transport/tcp/TCP.ned
	src/transport/tcp_nsc/TCP_NSC.h

+++ src/linklayer/ext/ExtInterface.h
@@ -46,9 +46,9 @@ class ExtInterface : public cSimpleModule
     InterfaceEntry *interfaceEntry;  // points into RoutingTable
 
     // statistics
-    uint64 numSent;
-    uint64 numRcvd;
-    uint64 numDropped;
+    int numSent;
+    int numRcvd;
+    int numDropped;
 
     // access to real network interface via Scheduler class:
     cSocketRTScheduler *rtScheduler;
